[RE] Fix shutdown sequence

+++ data_feed/perf/metrics/prom_connection.py
@@ -7,8 +7,10 @@ from perf.defines import PROM_PORT_FORWARD, PROM_POD_NAME, PROM_NAMESPACE
 class PromConnection:
     def __init__(self):
         self.forward_prom_port_proc = None
+        self.running = False
 
     def start(self):
+        self.running = True
         print(f'[PromConnection] Forwarding Prometheus port {PROM_PORT_FORWARD}...')
         # TODO check success of Popen
         self.forward_prom_port_proc = subprocess.Popen(
@@ -16,13 +18,20 @@ class PromConnection:
             shell=True,
             stdout=subprocess.DEVNULL,
         )
+        if not self.running:
+            return
         # 5s to spin up
         wait = 5
         print(f'[PromConnection] Waiting {wait}s to spin up Prometheus connection...')
         time.sleep(wait)
+        if not self.running:
+            return
         print(f'[PromConnection] Prometheus connection started on pid {self.forward_prom_port_proc.pid}')
 
     def stop(self):
+        if not self.running:
+            return
+        self.running = False
         if self.forward_prom_port_proc is not None:
             self.forward_prom_port_proc.terminate()
             self.forward_prom_port_proc.wait()
